requirements.txt
python-decouple
paho-mqtt
sentry-sdk>=0.14.3
